refactor: 도메인 검증 로직 validator로 분리
